% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{Add_Deconvolution_Info}
\alias{Add_Deconvolution_Info}
\title{Add RCTD results and weights to the barcodes tibble}
\usage{
Add_Deconvolution_Info(barcodes, deconv, addWeights = FALSE)
}
\arguments{
\item{barcodes}{The barcodes tibble}

\item{deconv}{The loaded RCTD deconvolution object}

\item{addWeights}{Add modified doublet weights to the tibble}
}
\value{
A \code{barcodes} tibble with the deconvolution results added to it
}
\description{
This function adds the RCTD deconvolution results to the previously generated
barcodes tibble.
}
\examples{
\dontrun{
Barcodes=Add_Deconvolution_Info(barcodes=Barcodes,deconv=Deconvolution)
}

}
\concept{deconvolution}
