% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{Add_Expression}
\alias{Add_Expression}
\title{Add gene expression data to the barcodes tibble}
\usage{
Add_Expression(barcodes, seurat, genes)
}
\arguments{
\item{barcodes}{The barcodes tibble}

\item{seurat}{The Seurat object of Visium HD analysis}

\item{genes}{A list of genes whose expression is to be added}
}
\value{
A \code{barcodes} tibble with the gene expression data added to it
}
\description{
This function uses a Seurat object to load specific gene expression data to
the barcodes tibble.
}
\examples{
\dontrun{
Barcodes=Add_Expression(barcodes=Barcodes,seurat=Seurat)
}

}
\concept{expression}
