% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{Compute_Cell_Distance}
\alias{Compute_Cell_Distance}
\title{Compute cell-cell distance based on the spatial coordinates}
\usage{
Compute_Cell_Distance(
  coordinates,
  interaction.range = NULL,
  ratio = NULL,
  tol = NULL
)
}
\arguments{
\item{coordinates}{a dataframe in which each row gives the spatial locations/coordinates of each cell/spot}

\item{interaction.range}{The maximum interaction/diffusion range of ligands. This hard threshold is used to filter out the connections between spatially distant cells}

\item{ratio}{The conversion factor when converting spatial coordinates from Pixels or other units to Micrometers (i.e.,Microns).

For example, setting `ratio = 0.25` indicates that 1 pixel equals 0.25um in the coordinates.
For 10X Visium HD, it is the ratio of the theoretical spot size (i.e., 8um) over the number of pixels that span the diameter of a theoretical spot size in the full-resolution image (i.e., 'spot_diameter_fullres' in the 'scalefactors_json.json' file).}

\item{tol}{The tolerance factor to increase the robustness when comparing the center-to-center distance against the `interaction.range`. This can be the half value of cell/spot size in the unit of um.

For example, for 10X visium, `tol` can be set as `8/2`; for slide-seq, `tol` can be set as `10/2`.
If the cell/spot size is not known, we provide a function `computeCellDistance` to compute the center-to-center distance. `tol` can be the the half value of the minimum center-to-center distance.}
}
\value{
an object of class "dist" giving the pairwise cell-cell distance
}
\description{
This function creates a large distance matrix d.spatial by calculating the
euclidean distance of each cell/spot from every other cell/spot
}
\examples{
\dontrun{
Distance <- Compute_Cell_Distance(coordinates = spatial.locs, ratio = spatial.factors$ratio, tol = spatial.factors$tol)
}

}
\concept{utility}
