% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Create_Images}
\alias{Create_Images}
\title{Create a tibble for plotting image}
\usage{
Create_Images(data.dir)
}
\arguments{
\item{data.dir}{Path to the directory containing the spatial directory}
}
\value{
An images tibble
}
\description{
This function creates a tibble: images that can be used to create
a graphical object (grob) that represents a raster image.
}
\examples{
\dontrun{
data_dir <- 'path/to/outs/directory'
list.files(data_dir) # Should show spatial folder
images=Create_Images(data.dir = data_dir)
}

}
\concept{preprocessing}
