% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Geom_Spatial}
\alias{Geom_Spatial}
\title{Visualize HE image on a spatial plot}
\usage{
Geom_Spatial(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = FALSE,
  ...
)
}
\arguments{
\item{barcodes}{The barcodes tibble}

\item{gene}{The gene whose expression is to be plotted}

\item{pt.size}{The dize of the points}

\item{shape}{The shape of the slice involved (either "circle" or "square")}
}
\value{
A list of ggplot objects
}
\description{
This function creates a graphical object (grob) that represents a raster
image from the images tibble
}
\examples{
\dontrun{
PlotA=Geom_Spatial(data=Images, , aes(grob = grob), x = 0.5, y = 0.5)
}

}
\concept{visualization}
