% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{Get_Circle}
\alias{Get_Circle}
\title{Create a character vector of a circular region for sub-setting}
\usage{
Get_Circle(
  spot,
  size_microns,
  barcodes,
  data.dir,
  celltype = NA,
  binsize = "008um"
)
}
\arguments{
\item{spot}{Barcode at the center of the circle to be generated}

\item{size_microns}{The radius of the circle in microns}

\item{barcodes}{The barcodes tibble}

\item{data.dir}{Path to the directory containing the binned_outputs directory}

\item{celltype}{The type of cell at the center}

\item{binsize}{The binsize of the current analysis (default="008um")}
}
\value{
A character vector of the barcodes in the circular region
}
\description{
This function uses a given radius size (in microns) whose center is a given
barcode to create a character vector of barcodes from the barcodes tibble
that can be used to subset a circular region of the Visium HD dataset.
}
\examples{
\dontrun{
data_dir <- 'path/to/outs/directory'
list.files(data_dir) # Should show binned_outputs folder
CircleA=Get_Circle(spot='s_008um_00128_00230-1', size_microns=350,
                    barcodes=Barcodes, data.dir=data_dir, celltype=NA,
                    binsize="008um")
}

}
\concept{slicing}
