% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{Get_Doublet_Weights_Modified}
\alias{Get_Doublet_Weights_Modified}
\title{Create a Weights matrix from RCTD results}
\usage{
Get_Doublet_Weights_Modified(results_df, weights, celltypes)
}
\arguments{
\item{results_df}{The RCTD results tibble}

\item{weights}{The doublet weights stored in RCTD}

\item{celltypes}{The cell types of the reference}
}
\value{
A modified \code{Weights} matrix
}
\description{
This function creates a modified weights matrix using the barcodes,
cell_types and doublet_weights.
}
\examples{
\dontrun{
Weights=Get_Doublet_Weights_Modified(results_df,weights,celltypes)
head(Weights)
}

}
\concept{deconvolution}
