% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{Get_Gene_Periphery}
\alias{Get_Gene_Periphery}
\title{Create a character vector of peripheral cells of a gene based niche for
sub-setting}
\usage{
Get_Gene_Periphery(
  barcodes,
  gene,
  distance = 50,
  data.dir,
  binsize = "008um",
  seurat
)
}
\arguments{
\item{barcodes}{The barcodes tibble}

\item{gene}{The gene at the center}

\item{data.dir}{Path to the directory containing the binned_outputs directory}

\item{binsize}{The binsize of the current analysis (default="008um")}

\item{seurat}{The Seurat object of Visium HD analysis}

\item{spot}{Barcode at the center of the rectangle to be generated}

\item{size_microns}{The radius of the circle in microns}
}
\value{
A character vector of the barcodes surrounding a cell expressing a
gene
}
\description{
This function uses a given size (in microns) whose center is a cell
expressing a particualr gene to create a character vector of barcodes from
the barcodes tibble that can be used to subset peripheral cells of a niche
from the Visium HD dataset.
}
\examples{
\dontrun{
data_dir <- 'path/to/outs/directory'
list.files(data_dir) # Should show binned_outputs folder
PeripheryA=Get_Gene_Periphery(barcodes=Barcodes, gene="GAPDH", distance= 50,
                    data.dir = data_dir, binsize="008um", seurat=Seurat)
}

}
\concept{slicing}
