% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{Get_Periphery}
\alias{Get_Periphery}
\title{Create a character vector of peripheral cells of a cell based niche for
sub-setting}
\usage{
Get_Periphery(barcodes, celltype, distance = 50, data.dir, binsize = "008um")
}
\arguments{
\item{barcodes}{The barcodes tibble}

\item{celltype}{The type of cell at the center}

\item{data.dir}{Path to the directory containing the binned_outputs directory}

\item{binsize}{The binsize of the current analysis (default="008um")}

\item{spot}{Barcode at the center of the rectangle to be generated}

\item{size_microns}{The radius of the circle in microns}
}
\value{
A character vector of the barcodes surrounding a cell type
}
\description{
This function uses a given size (in microns) whose center is a given cell
type to create a character vector of barcodes from the barcodes tibble that
can be used to subset peripheral cells of a niche from the Visium HD dataset.
}
\examples{
\dontrun{
data_dir <- 'path/to/outs/directory'
list.files(data_dir) # Should show binned_outputs folder
PeripheryA=Get_Periphery(barcodes=Barcodes, celltype=Cell_Type, distance= 50,
                    data.dir = data_dir, binsize="008um")
}

}
\concept{slicing}
