% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{Get_Rectangle}
\alias{Get_Rectangle}
\title{Create a character vector of a rectangular region for sub-setting}
\usage{
Get_Rectangle(spot, size_microns, xfactor, yfactor, barcodes, binsize = 8)
}
\arguments{
\item{spot}{Barcode at the center of the rectangle to be generated}

\item{size_microns}{The length of the rectangle in microns}

\item{xfactor}{The factor of the horizontal line of the rectangle}

\item{yfactor}{The factor of the vertical line of the rectangle}

\item{barcodes}{The barcodes tibble}

\item{binsize}{The binsize of the current analysis (default=8)}
}
\value{
A character vector of the barcodes in the rectangular region
}
\description{
This function uses a given size (in microns) whose center is a given barcode
to create a character vector of barcodes from the barcodes tibble that can be
used to subset a rectangular region of the Visium HD dataset.
}
\examples{
\dontrun{
RectangleA=Get_Rectangle(spot='s_008um_00128_00230-1', size_microns=350,
                    barcodes=Barcodes, binsize=8)
}

}
\concept{slicing}
