% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Get_Spatial_Files}
\alias{Get_Spatial_Files}
\title{Create paths and read different spatial files found in outs directory}
\usage{
Get_Spatial_Files(data.dir, type)
}
\arguments{
\item{data.dir}{Path to the directory containing the spatial directory}

\item{type}{Specifies the type of file required}
}
\value{
A bitmap image
}
\description{
This function generates paths to the appropriate directory and creates all
the spatial files.
}
\examples{
\dontrun{
data_dir <- 'path/to/outs/directory'
list.files(data_dir) # Should show spatial folder
Image=Get_Spatial_Files(data.dir = data_dir, type="tissue_lowres_image")
}

}
\concept{preprocessing}
