% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slicing.R
\name{Get_Square}
\alias{Get_Square}
\title{Create a character vector of a square region for sub-setting}
\usage{
Get_Square(spot, size_microns, barcodes, binsize = 8)
}
\arguments{
\item{spot}{Barcode at the center of the square to be generated}

\item{size_microns}{The length of the square in microns}

\item{barcodes}{The barcodes tibble}

\item{binsize}{The binsize of the current analysis (default=8)}
}
\value{
A character vector of the barcodes in the square region
}
\description{
This function uses a given size (in microns) whose center is a given barcode
to create a character vector of barcodes from the barcodes tibble that can be
used to subset a square region of the Visium HD dataset.
}
\examples{
\dontrun{
SquareA=Get_Square(spot='s_008um_00128_00230-1', size_microns=350,
                    barcodes=Barcodes, binsize=8)
}

}
\concept{slicing}
