% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{\%!in\%}
\alias{\%!in\%}
\title{The inverse %in% operator}
\usage{
x \%!in\% y
}
\arguments{
\item{x}{vector}

\item{y}{vector of same type as x}
}
\value{
A logical vector
}
\description{
Matching operator (x not in y) oposed to the %in%-operator (x in y)
}

\concept{utility}
