% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{Load_RCTD}
\alias{Load_RCTD}
\title{Load a RCTD result for deconvolution}
\usage{
Load_RCTD(data)
}
\arguments{
\item{data}{A sapcexr RCTD result rds object}
}
\value{
A list with \code{RCTD_df} tibble and \code{Weights} matrix
}
\description{
This function loads a sapcexr RCTD result rds object with the appropriate
weights that can be used for deconvolution 10X Visium HD dataset.
}
\examples{
\dontrun{
data <- 'path/to/RCTD_rds/directory'
Deconvolution=Load_RCTD(data = data)
}

}
\concept{deconvolution}
