% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{Load_Spatial_Data}
\alias{Load_Spatial_Data}
\title{Load a 10X Visium HD dataset for niche detection}
\usage{
Load_Spatial_Data(data.dir, size = "008um")
}
\arguments{
\item{data.dir}{Path to the directory containing the binned_outputs directory}

\item{size}{Specifies the bin sizes to read in - defaults to 8um bin}
}
\value{
A list with two tibbles: \code{barcodes} and \code{images}
}
\description{
This function loads a Visium HD output directory and generate a list
containing two tibbles: barcodes with all bins and images for plotting.
}
\examples{
\dontrun{
data_dir <- 'path/to/outs/directory'
list.files(data_dir) # Should show binned_outputs folder
VisiumHD=Load_Spatial_Data(data.dir = data_dir)
Barcodes=VisiumHD$barcodes
}

}
\concept{preprocessing}
