% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{Spatial_Feature_Plot}
\alias{Spatial_Feature_Plot}
\title{Visualize 'features' i.e. genes on a spatial plot}
\usage{
Spatial_Feature_Plot(barcodes, gene, pt.size = 2, shape = "circle")
}
\arguments{
\item{barcodes}{The barcodes tibble}

\item{gene}{The gene whose expression is to be plotted}

\item{pt.size}{The dize of the points}

\item{shape}{The shape of the slice involved (either "circle" or "square")}
}
\value{
A list of ggplot objects
}
\description{
This function colors bins on spatial plot according to a 'feature'
(i.e. gene expression, PC scores, number of genes detected, etc.)
}
\examples{
\dontrun{
PlotA=Spatial_Feature_Plot(barcodes=Barcodes, gene="GAPDH", pt.size=2,
                            shape="circle")
}

}
\concept{visualization}
