% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expression.R
\name{Add_Expression}
\alias{Add_Expression}
\title{Add gene expression data to the barcodes tibble}
\usage{
Add_Expression(barcodes, seurat, gene)
}
\arguments{
\item{barcodes}{The barcodes tibble}

\item{seurat}{The Seurat object of Visium HD analysis}

\item{gene}{The gene at the center}
}
\value{
A \code{barcodes} tibble with the gene expression data added to it
}
\description{
This function adds gene expression data from a Seurat object to the
previously generated barcodes tibble.
}
\examples{
\dontrun{
Barcodes=Add_Expression(barcodes=Barcodes, gene="GAPDH", seurat=Seurat)
}

}
\concept{expression}
